Changed resolution numbers to integers

+++ frameinfo.h
@@ -4,7 +4,7 @@ class frameinfo
 {
 public:
 	double winl, winr, winb, wint;
-	long resx, resy;
+	int resx, resy;
 	double deltax, deltay;
 	long iters;
 };
